refactor: Consolidate enhanced bootloader service functionality into IBootloaderService and standardize memory region path handling.
