+++ benchmarks/S7Tools.Benchmarks/ProfileCrudBenchmarks.cs
@@ -137,7 +137,6 @@ public class ProfileCrudBenchmarks
         public string PowerSupplyProfilesPath => Path.Combine(ProfilesDirectory, "PowerSupply", "PowerSupplyProfiles.json");
         public string MemoryRegionProfilesPath => Path.Combine(ProfilesDirectory, "MemoryRegions", "MemoryRegionProfiles.json");
         public string PayloadSetProfilesPath => Path.Combine(ProfilesDirectory, "PayloadSets", "PayloadSetProfiles.json");
-        public string MemoryRegionsDirectory => Path.Combine(ResourcesDirectory, "MemoryRegions");
         public string LogsDirectory => Path.Combine(ResourcesDirectory, "Logs");
         public string MainLogsDirectory => Path.Combine(LogsDirectory, "Main");
         public string ExportedLogsDirectory => Path.Combine(LogsDirectory, "Exported");
